3.3-permissions/api_with_restrictions/advertisements/views.py
from rest_framework.exceptions import PermissionDenied
from rest_framework.permissions import IsAuthenticated
from rest_framework.viewsets import ModelViewSet
from .models import Advertisement
from .serializers import AdvertisementSerializer
from .filters import AdvertisementFilter

class AdvertisementViewSet(ModelViewSet):
    """ViewSet для объявлений."""

    queryset = Advertisement.objects.all()
    serializer_class = AdvertisementSerializer
    filterset_class = AdvertisementFilter  # Подключение фильтров

    def get_permissions(self):
        """Получение прав для действий."""
        if self.action in ["create", "update", "partial_update"]:
            return [IsAuthenticated()]
        return []

    def perform_update(self, serializer):
        """Метод для выполнения обновления."""
        advertisement = self.get_object()
        if advertisement.creator != self.request.user:
            raise PermissionDenied("Вы не можете изменять это объявление.")
        serializer.save()
