3.3-permissions/api_with_restrictions/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-filter==24.3
